feat: impedir exclusão de categorias com despesas associadas

+++ src/main/java/com/example/despesas/services/CategoriaService.java
@@ -2,8 +2,11 @@ package com.example.despesas.services;
 
 import com.example.despesas.entities.Categoria;
 import com.example.despesas.repositories.CategoriaRepository;
+import com.example.despesas.repositories.DespesaRepository;
+import com.example.despesas.services.exceptions.DataBaseException;
 import com.example.despesas.services.exceptions.ResourceNotFoundException;
 import jakarta.transaction.Transactional;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,9 +15,11 @@ import java.util.List;
 public class CategoriaService {
 
     private final CategoriaRepository categoriaRepository;
+    private final DespesaRepository despesaRepository;
 
-    public CategoriaService(CategoriaRepository categoriaRepository) {
+    public CategoriaService(CategoriaRepository categoriaRepository, DespesaRepository despesaRepository) {
         this.categoriaRepository = categoriaRepository;
+        this.despesaRepository = despesaRepository;
     }
 
     public List<Categoria> listarCategorias() {
@@ -38,11 +43,18 @@ public class CategoriaService {
 
     @Transactional
     public void deletarCategoria(Long id) {
-        Categoria obj = buscarCategoriaPorId(id);
-        categoriaRepository.delete(obj);
+            Categoria obj = buscarCategoriaPorId(id);
+            if(despesaRepository.existsByCategoria_Id(obj.getId())) {
+                throw new DataBaseException("Não é possível excluir uma categoria que possui despesas associadas.");
+            }
+            categoriaRepository.delete(obj);
     }
 
     private void atualizarDados(Categoria obj, Categoria novosDados) {
         obj.setNome(novosDados.getNome());
+    }
+
+    public boolean possuiDespesas(Long id) {
+        return despesaRepository.existsByCategoria_Id(id);
     }
 }
